Correct printf usage in test case

+++ tests/workgroup/print_all_ids.cl
@@ -1,11 +1,10 @@
 kernel void test_kernel() {
-  printf("global: %zd-%zd-%zd\n",
-           get_global_id(0),
-           get_global_id(1),
-           get_global_id(2));
-  printf("local: %zd-%zd-%zd\n",
-           get_local_id(0),
-           get_local_id(1),
-           get_local_id(2));
+  printf("global: %d-%d-%d\n",
+           (int)get_global_id(0),
+           (int)get_global_id(1),
+           (int)get_global_id(2));
+  printf("local: %d-%d-%d\n",
+           (int)get_local_id(0),
+           (int)get_local_id(1),
+           (int)get_local_id(2));
 }
-
